refactor: create EventType interface and use it in EventType classes
