+++ privmx-endpoint/src/main/java/com/simplito/java/privmx_endpoint/model/events/eventTypes/EventType.java
@@ -1,0 +1,3 @@
+package com.simplito.java.privmx_endpoint.model.events.eventTypes;
+
+public interface EventType { }
