+++ privmx-endpoint/src/main/java/com/simplito/java/privmx_endpoint/model/events/eventTypes/InboxEventType.java
@@ -1,0 +1,9 @@
+package com.simplito.java.privmx_endpoint.model.events.eventTypes;
+
+public enum InboxEventType implements EventType {
+    INBOX_CREATE,
+    INBOX_UPDATE,
+    INBOX_DELETE,
+    ENTRY_CREATE,
+    ENTRY_DELETE
+}
