+++ privmx-endpoint/src/main/java/com/simplito/java/privmx_endpoint/model/events/eventTypes/KvdbEventType.java
@@ -1,6 +1,6 @@
-package com.simplito.java.privmx_endpoint.model;
+package com.simplito.java.privmx_endpoint.model.events.eventTypes;
 
-public enum KvdbEventType {
+public enum KvdbEventType implements EventType {
     KVDB_CREATE,
     KVDB_UPDATE,
     KVDB_DELETE,
